src/utils/image_processor.py
"""Image processing module with AI-powered enhancement."""

import logging
import torch
import torch.nn as nn
import torch.nn.functional as F
import numpy as np
from PIL import Image, ImageEnhance
from enum import Enum
from typing import Optional, Tuple, List, Dict
import cv2
import torchvision.transforms as transforms
from torchvision.models import resnet18
import time
import os

# Suppress torch warnings
os.environ['PYTORCH_JIT'] = '0'
torch.hub.set_dir(os.path.expanduser('~/.cache/torch/hub'))

logger = logging.getLogger(__name__)

class EnhancementStrategy(Enum):
    AUTO = "auto"
    BALANCED = "balanced"
    SHARPNESS = "sharpness"
    CONTRAST = "contrast"
    COLOR = "color"
    DETAIL = "detail"
    NOISE_REDUCTION = "noise_reduction"

class QualityNet(nn.Module):
    """Neural network for image quality assessment."""
    def __init__(self):
        super().__init__()
        # Use pretrained ResNet as base
        self.base = resnet18(weights='IMAGENET1K_V1')
        # Modify for quality assessment
        self.base.fc = nn.Sequential(
            nn.Linear(512, 256),
            nn.ReLU(),
            nn.Dropout(0.5),
            nn.Linear(256, 5)  # 5 quality metrics
        )
        
        # Initialize new layers
        for m in self.base.fc.modules():
            if isinstance(m, nn.Linear):
                nn.init.kaiming_normal_(m.weight)
                nn.init.constant_(m.bias, 0)
        
    def forward(self, x):
        return torch.sigmoid(self.base(x))

class ImageProcessor:
    def __init__(self):
        """Initialize the image processor."""
        self.sharpness_threshold = 0.95  # Increased for maximum detail
        self.contrast_threshold = 0.90   # Optimized for dynamic range
        self.noise_threshold = 0.15      # Lowered to preserve fine details
        self.detail_threshold = 0.95     # Increased for enhanced clarity
        self.color_threshold = 0.85      # Balanced for natural look
        self._enhancement_history = []
        
        # Initialize quality assessment model
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.quality_net = QualityNet().to(self.device)
        
        # Initialize transforms for 5K resolution
        self.transform = transforms.Compose([
            transforms.Resize((5120, 2880)),  # 5K resolution
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.485, 0.456, 0.406],
                              std=[0.229, 0.224, 0.225])
        ])
        
        # Set model to eval mode
        self.quality_net.eval()
        
        logger.info(f"Initialized image processor with device: {self.device}")

    def assess_quality(self, image: Image.Image) -> Dict[str, float]:
        """Assess image quality using neural network."""
        with torch.no_grad():
            # Transform image
            x = self.transform(image).unsqueeze(0).to(self.device)
            # Get quality scores
            scores = self.quality_net(x)[0].cpu().numpy()
            return {
                'sharpness': float(scores[0]),
                'contrast': float(scores[1]),
                'noise_level': float(scores[2]),
                'detail': float(scores[3]),
                'color': float(scores[4])
            }

    def _apply_clahe(self, image: np.ndarray) -> np.ndarray:
        """Apply CLAHE while preserving color."""
        # Convert to LAB
        lab = cv2.cvtColor(image, cv2.COLOR_RGB2LAB)
        l, a, b = cv2.split(lab)
        
        # Apply CLAHE to L channel with optimized parameters
        clahe = cv2.createCLAHE(clipLimit=3.0, tileGridSize=(16,16))
        l = clahe.apply(l)
        
        # Merge back
        lab = cv2.merge([l, a, b])
        return cv2.cvtColor(lab, cv2.COLOR_LAB2RGB)

    def enhance_image(self, image: Image.Image, strategy: Optional[EnhancementStrategy] = None, 
                     enhancement_params: Optional[Dict] = None) -> Image.Image:
        """Enhance image using traditional methods with multiple attempts."""
        try:
            if not isinstance(image, Image.Image):
                raise ValueError("Input must be a PIL Image")

            # Convert to RGB
            image = image.convert('RGB')
            
            # Resize to 5K resolution
            image = image.resize((5120, 2880), Image.Resampling.LANCZOS)
            
            # Get enhancement parameters
            params = enhancement_params or {
                'contrast': 1.4,    # Increased for better dynamic range
                'sharpness': 1.5,   # Maximum sharpness
                'color': 1.2        # Enhanced color
            }
            
            # Create copy for enhancement
            enhanced = image.copy()
            
            if strategy in [EnhancementStrategy.AUTO, EnhancementStrategy.BALANCED, None]:
                # Convert to numpy for CLAHE
                img_array = np.array(enhanced)
                img_array = self._apply_clahe(img_array)
                enhanced = Image.fromarray(img_array)
                
                # Enhanced contrast
                enhancer = ImageEnhance.Contrast(enhanced)
                enhanced = enhancer.enhance(params['contrast'])
                
                # Maximum sharpness
                enhancer = ImageEnhance.Sharpness(enhanced)
                enhanced = enhancer.enhance(params['sharpness'])
                
                # Optimized color
                enhancer = ImageEnhance.Color(enhanced)
                enhanced = enhancer.enhance(params['color'])
                
                # Apply unsharp masking for extra detail
                img_array = np.array(enhanced)
                gaussian = cv2.GaussianBlur(img_array, (0, 0), 3.0)
                enhanced = Image.fromarray(
                    cv2.addWeighted(img_array, 1.8, gaussian, -0.8, 0)
                )
                
            elif strategy == EnhancementStrategy.SHARPNESS:
                enhancer = ImageEnhance.Sharpness(enhanced)
                enhanced = enhancer.enhance(params['sharpness'] * 1.2)
                
            elif strategy == EnhancementStrategy.CONTRAST:
                # Apply CLAHE
                img_array = np.array(enhanced)
                img_array = self._apply_clahe(img_array)
                enhanced = Image.fromarray(img_array)
                
                # Additional contrast enhancement
                enhancer = ImageEnhance.Contrast(enhanced)
                enhanced = enhancer.enhance(params['contrast'])
                
            elif strategy == EnhancementStrategy.COLOR:
                enhancer = ImageEnhance.Color(enhanced)
                enhanced = enhancer.enhance(params['color'])
                
            elif strategy == EnhancementStrategy.DETAIL:
                # Apply advanced unsharp masking
                img_array = np.array(enhanced)
                gaussian = cv2.GaussianBlur(img_array, (0, 0), 3.0)
                enhanced = Image.fromarray(
                    cv2.addWeighted(img_array, 1.8, gaussian, -0.8, 0)
                )
                
            elif strategy == EnhancementStrategy.NOISE_REDUCTION:
                enhanced = Image.fromarray(
                    cv2.fastNlMeansDenoisingColored(
                        np.array(enhanced), 
                        None, 
                        10, 10, 7, 21
                    )
                )
            
            return enhanced
            
        except Exception as e:
            logger.error(f"Error enhancing image: {str(e)}")
            return image

    def calculate_metrics(self, image: Image.Image) -> dict:
        """Calculate quality metrics using AI model."""
        return self.assess_quality(image)

    def get_enhancement_history(self) -> List[dict]:
        """Get the history of enhancement attempts."""
        return self._enhancement_history

    def update_thresholds(self, **kwargs):
        """Update the enhancement thresholds."""
        if 'sharpness' in kwargs:
            self.sharpness_threshold = kwargs['sharpness']
        if 'contrast' in kwargs:
            self.contrast_threshold = kwargs['contrast']
        if 'noise' in kwargs:
            self.noise_threshold = kwargs['noise']
        if 'detail' in kwargs:
            self.detail_threshold = kwargs['detail']
        if 'color' in kwargs:
            self.color_threshold = kwargs['color']
